+++ changelog.md
@@ -1,3 +1,5 @@
+ - EVOL: expose config file path (py side) : self.cfg._file
+
 0.5.3: (11/02/2020)
  - FIX: thanks @icarito https://github.com/manatlan/guy/pull/8/commits/9477b548b91ae81bfc327dac7ba1ec80804f4f8d